# medical-project-backend/backend/app/static/openapi.yaml
openapi: 3.1.0
info:
  title: Patient Management API
  version: "1.2.0"
  description: >
    API for managing patients, including creating new patients, updating existing patient records (with version history), medical history, medical visits, and retrieving detailed patient information.
servers:
  - url: https://8dj6t53su8wv.share.zrok.io
    description: Local development server
...
paths:
...
  /api/users/{user_id}:
    get:
      security:
        - BearerAuth: []
      tags:
        - Users
      summary: Retrieve a User
      description: Retrieves detailed information about a user identified by their user_id.
      parameters:
        - in: path
          name: user_id
          required: true
          schema:
            type: string
            format: uuid
          description: The unique identifier of the user to retrieve.
      responses:
        '200':
          description: User retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    put:
      security:
        - BearerAuth: []
      tags:
        - Users
      summary: Update a User
      description: Updates an existing user's information while preserving the old version in history.
      parameters:
        - in: path
          name: user_id
          required: true
          schema:
            type: string
            format: uuid
          description: The unique identifier of the user to update.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserUpdate'
            examples:
              Example1:
                summary: Sample User Update
                value:
                  login: "johndoe"
                  first_name: "John"
                  last_name: "Doe"
                  phone_number: "123-456-7890"
                  email: "john.doe@example.com"
      responses:
        '201':
          description: User updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserUpdateResponse'
        '400':
          description: Bad Request - Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      security:
        - BearerAuth: []
      tags:
        - Users
      summary: Hide a User
      description: >
        Makes a user hidden in the system without actually deleting them.
        This preserves data integrity while removing them from standard queries.
      parameters:
        - in: path
          name: user_id
          required: true
          schema:
            type: string
            format: uuid
          description: The unique identifier of the user to hide.
      responses:
        '200':
          description: User successfully hidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusResponse'
              examples:
                SuccessResponse:
                  summary: Successful Hiding
                  value:
                    status: "success"
        '403':
          description: Forbidden - Insufficient permissions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/users/{user_id}/credentials:
    put:
      security:
        - BearerAuth: []
      tags:
        - Users
      summary: Update User Credentials
      description: >
        Updates a user's login and/or password information.
        This requires appropriate permissions (admin, doctor, or healthcare professional).
      parameters:
        - in: path
          name: user_id
          required: true
          schema:
            type: string
            format: uuid
          description: The unique identifier of the user whose credentials are being updated.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CredentialsUpdate'
            examples:
              Example1:
                summary: Credential Update Example
                value:
                  login: "newlogin"
                  password: "newpassword123"
      responses:
        '201':
          description: Credentials updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusResponse'
              examples:
                SuccessResponse:
                  summary: Successful Update
                  value:
                    status: "success"
        '400':
          description: Bad Request - Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden - Insufficient permissions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Not Found - User does not exist
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from /api/auth/login
  schemas:
...
    UserResponse:
      type: object
      description: Schema for user details.
      properties:
        user_id:
          type: string
          format: uuid
          example: "99febb18-8a4c-40b6-942f-df3c60d522dd"
        login:
          type: string
          example: "parent1"
        user_type:
          type: string
          example: "PATIENT"
        first_name:
          type: string
          example: "Parent"
        last_name:
          type: string
          example: "One"
        phone_number:
          type: string
          example: "111-222-3333"
        email:
          type: string
          format: email
          example: "parent1@example.com"
        created_at:
          type: string
          format: date-time
          example: "Fri, 14 Feb 2025 09:33:35 GMT"
        modified_at:
          type: string
          format: date-time
          example: "Fri, 14 Feb 2025 09:33:35 GMT"
      required:
        - user_id
        - login
        - user_type
        - first_name
        - last_name
        - phone_number
        - email
        - created_at
        - modified_at
      additionalProperties: false
    ErrorResponse:
      type: object
      description: Schema for error responses.
      properties:
        error:
          type: string
          example: "Invalid foreign key reference."
      required:
        - error
      additionalProperties: false
...
    UserUpdate:
      type: object
      description: Schema for updating an existing user.
      properties:
        login:
          type: string
          example: "johndoe"
          description: Updated login identifier.
        first_name:
          type: string
          example: "John"
          description: Updated first name.
        last_name:
          type: string
          example: "Doe"
          description: Updated last name.
        phone_number:
          type: string
          example: "123-456-7890"
          description: Updated phone number.
        email:
          type: string
          format: email
          example: "john.doe@example.com"
          description: Updated email address.
      required:
        - login
        - first_name
        - last_name
        - phone_number
        - email
      additionalProperties: false
    UserUpdateResponse:
      type: object
      description: Response schema for an updated user.
      properties:
        user_id:
          type: string
          format: uuid
          example: "a1b2c3d4-e5f6-7890-abcd-ef1234567890"
        login:
          type: string
          example: "johndoe"
        user_type:
          type: string
          example: "PATIENT"
        first_name:
          type: string
          example: "John"
        last_name:
          type: string
          example: "Doe"
        phone_number:
          type: string
          example: "123-456-7890"
        email:
          type: string
          format: email
          example: "john.doe@example.com"
        created_at:
          type: string
          format: date-time
          example: "Fri, 14 Feb 2025 09:33:35 GMT"
      required:
        - user_id
        - login
        - user_type
        - first_name
        - last_name
        - phone_number
        - email
        - created_at
      additionalProperties: false
...
    CredentialsUpdate:
      type: object
      description: Schema for updating user credentials.
      properties:
        login:
          type: string
          example: "johndoe"
          description: Updated login identifier.
        password:
          type: string
          example: "MySecretPass123"
          description: New password in plain text.
      required:
        - login
        - password
      additionalProperties: false
...
    StatusResponse:
      type: object
      description: Simple status response.
      properties:
        status:
          type: string
          example: "success"
      required:
        - status
      additionalProperties: false
...